added automatic deletion of test account

+++ src/db_schema.sql
@@ -188,9 +188,9 @@ INSERT INTO Cities (city_id, city_owner, population, population_capacity, town) 
 DELIMITER &&
 CREATE PROCEDURE reset_tests ()
 BEGIN
-    DELETE FROM Buildings WHERE city_id=(SELECT city_id FROM Cities JOIN Accounts ON city_owner=player_id WHERE username='User200');
-    DELETE FROM Cities WHERE city_id=(SELECT city_id FROM (SELECT * FROM Cities) AS TempCities JOIN Accounts ON city_owner=player_id WHERE username='User200');
-    DELETE FROM Sessions WHERE player_id=(SELECT player_id FROM Accounts WHERE username='User200');
-    DELETE FROM Accounts WHERE username='User200';
+    DELETE FROM city_io.Buildings WHERE city_id=(SELECT city_id FROM city_io.Cities JOIN city_io.Accounts ON city_owner=player_id WHERE username='User200');
+    DELETE FROM city_io.Cities WHERE city_id=(SELECT city_id FROM (SELECT * FROM city_io.Cities) AS TempCities JOIN city_io.Accounts ON city_owner=player_id WHERE username='User200');
+    DELETE FROM city_io.Sessions WHERE player_id=(SELECT player_id FROM city_io.Accounts WHERE username='User200');
+    DELETE FROM city_io.Accounts WHERE username='User200';
 END &&
 DELIMITER ;
